+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.10.9
+
+This is a patch release that:
+
+- Fixes a bug that allowed users to access private declarations in other
+  libraries with dot shorthands. (issue [dart-lang/sdk#62504])
+
+[dart-lang/sdk#62504]: https://github.com/dart-lang/sdk/issues/62504
+
 ## 3.10.8
 
 **Released on:** 2026-01-27